requirements.txt
google-adk
google-adk[cli]
python-dotenv
litellm
pyyaml